requirements.txt
python-craigslist
slackclient
sqlalchemy
